fix(db): encrypt provider key fixtures
